requirements.txt
flask==3.0.2
python-dotenv==0.21.0
openai==1.9.0
pymilvus==2.3.6
numpy==1.26.3